requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
watchfiles==0.21.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
azure-storage-blob==12.23.0
aiohttp==3.10.11
mediatr==1.2.0
python-dotenv==1.0.1
dependency-injector==4.42.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
pytest-mock==3.14.0
aiosqlite==0.20.0
openpyxl==3.1.2
python-telegram-bot==21.9

